Remove dependencies to other presenters to RoomMembersModerationPresenter. Move canDisplayModerationActions from presenter API to the state it emits.
