+++ features/roomdetails/impl/src/main/kotlin/io/element/android/features/roomdetails/impl/members/moderation/RoomMembersModerationState.kt
@@ -13,6 +13,7 @@ import io.element.android.libraries.matrix.api.room.RoomMember
 import kotlinx.collections.immutable.ImmutableList
 
 data class RoomMembersModerationState(
+    val canDisplayModerationActions: Boolean,
     val selectedRoomMember: RoomMember?,
     val actions: ImmutableList<ModerationAction>,
     val kickUserAsyncAction: AsyncAction<Unit>,
