+++ features/roomdetails/impl/src/test/kotlin/io/element/android/features/roomdetails/members/RoomMemberListPresenterTest.kt
@@ -19,8 +19,8 @@ import io.element.android.features.roomdetails.impl.members.aRoomMemberList
 import io.element.android.features.roomdetails.impl.members.aVictor
 import io.element.android.features.roomdetails.impl.members.aWalter
 import io.element.android.features.roomdetails.impl.members.moderation.RoomMembersModerationEvents
+import io.element.android.features.roomdetails.impl.members.moderation.RoomMembersModerationState
 import io.element.android.features.roomdetails.impl.members.moderation.aRoomMembersModerationState
-import io.element.android.features.roomdetails.members.moderation.FakeRoomMembersModerationPresenter
 import io.element.android.libraries.core.coroutine.CoroutineDispatchers
 import io.element.android.libraries.designsystem.theme.components.SearchBarResultState
 import io.element.android.libraries.matrix.api.core.UserId
@@ -28,6 +28,7 @@ import io.element.android.libraries.matrix.api.room.MatrixRoom
 import io.element.android.libraries.matrix.api.room.MatrixRoomMembersState
 import io.element.android.libraries.matrix.test.room.FakeMatrixRoom
 import io.element.android.libraries.matrix.test.room.aRoomInfo
+import io.element.android.tests.testutils.EventsRecorder
 import io.element.android.tests.testutils.WarmUpRule
 import io.element.android.tests.testutils.testCoroutineDispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -194,9 +195,9 @@ class RoomMemberListPresenterTest {
     @Test
     fun `present - RoomMemberSelected by default opens the room member details through the navigator`() = runTest {
         val navigator = FakeRoomMemberListNavigator()
-        val moderationPresenter = FakeRoomMembersModerationPresenter(canDisplayModerationActions = false)
+        val roomMembersModerationStateLambda = { aRoomMembersModerationState(canDisplayModerationActions = false) }
         val presenter = createPresenter(
-            moderationPresenter = moderationPresenter,
+            roomMembersModerationStateLambda = roomMembersModerationStateLambda,
             navigator = navigator,
             matrixRoom = FakeMatrixRoom(
                 updateMembersResult = { Result.success(Unit) },
@@ -215,17 +216,15 @@ class RoomMemberListPresenterTest {
     @Test
     fun `present - RoomMemberSelected will open the moderation options if the current user can use them`() = runTest {
         val navigator = FakeRoomMemberListNavigator()
-        var selectRoomMemberCallCounts = 0
-        val capturingState = aRoomMembersModerationState(eventSink = { event ->
-            if (event is RoomMembersModerationEvents.SelectRoomMember) {
-                selectRoomMemberCallCounts++
-            }
-        })
-        val moderationPresenter = FakeRoomMembersModerationPresenter(canDisplayModerationActions = true).apply {
-            givenState(capturingState)
+        val eventsRecorder = EventsRecorder<RoomMembersModerationEvents>()
+        val roomMembersModerationStateLambda = {
+            aRoomMembersModerationState(
+                canDisplayModerationActions = true,
+                eventSink = eventsRecorder,
+            )
         }
         val presenter = createPresenter(
-            moderationPresenter = moderationPresenter,
+            roomMembersModerationStateLambda = roomMembersModerationStateLambda,
             navigator = navigator,
             matrixRoom = FakeMatrixRoom(
                 updateMembersResult = { Result.success(Unit) },
@@ -237,7 +236,7 @@ class RoomMemberListPresenterTest {
         }.test {
             skipItems(1)
             awaitItem().eventSink(RoomMemberListEvents.RoomMemberSelected(aVictor()))
-            assertThat(selectRoomMemberCallCounts).isEqualTo(1)
+            eventsRecorder.assertSingle(RoomMembersModerationEvents.SelectRoomMember(aVictor()))
         }
     }
 }
@@ -266,12 +265,12 @@ private fun TestScope.createPresenter(
         updateMembersResult = { Result.success(Unit) }
     ),
     roomMemberListDataSource: RoomMemberListDataSource = createDataSource(coroutineDispatchers = coroutineDispatchers),
-    moderationPresenter: FakeRoomMembersModerationPresenter = FakeRoomMembersModerationPresenter(),
+    roomMembersModerationStateLambda: () -> RoomMembersModerationState = { aRoomMembersModerationState() },
     navigator: RoomMemberListNavigator = object : RoomMemberListNavigator {}
 ) = RoomMemberListPresenter(
     room = matrixRoom,
     roomMemberListDataSource = roomMemberListDataSource,
     coroutineDispatchers = coroutineDispatchers,
-    roomMembersModerationPresenter = moderationPresenter,
+    roomMembersModerationPresenter = { roomMembersModerationStateLambda() },
     navigator = navigator
 )
